refactor: generation 모듈 파일 분리
